STYLE: rename namic_dicom_typing file to utility_functions
